Add !timestamp tag to get timestamps in a mule yaml
I also renamed the file to be a general mule loader, instead of just a yaml loader for env vars
